+++ scripts/gui/ranking.gd
@@ -4,6 +4,8 @@ class_name Ranking
 extends VBoxContainer
 
 var _first_marble: Marble = null
+var _second_marble: Marble = null
+var _third_marble: Marble = null
 var _last_marble: Marble = null
 
 
@@ -19,6 +21,10 @@ func update() -> void:
 
 	if len(arr) > 0:
 		_first_marble = arr[0].get_marble()
+	if len(arr) > 1:
+		_second_marble = arr[1].get_marble()
+	if len(arr) > 2:
+		_third_marble = arr[2].get_marble()
 
 	var rank := 0
 	for child in arr:
@@ -36,24 +42,27 @@ func more_checkpoint(a: Participant, b: Participant) -> bool:
 	var b_marble = b.get_marble()
 
 	var out: bool
-	if (
-		a_marble.has_finish() and b_marble.has_finish()
-		or a_marble.get_checkpoint_count() == b_marble.get_checkpoint_count()
-	):
+	if a_marble.has_finish() and b_marble.has_finish():
 		out = a.get_rank() < b.get_rank()
 	elif a_marble.has_finish():
 		out = true
 	elif b_marble.has_finish():
 		out = false
 	else:
-		if a_marble.has_explode() and b_marble.has_explode():
+		if (
+			(a_marble.has_explode() or a_marble.has_oob())
+			and (b_marble.has_explode() or b_marble.has_oob())
+		):
 			out = a.get_rank() < b.get_rank()
-		elif a_marble.has_explode():
+		elif a_marble.has_explode() or a_marble.has_oob():
 			out = false
-		elif b_marble.has_explode():
+		elif b_marble.has_explode() or b_marble.has_oob():
 			out = true
 		else:
-			out = a_marble.get_checkpoint_count() > b_marble.get_checkpoint_count()
+			if a_marble.get_checkpoint_count() == b_marble.get_checkpoint_count():
+				out = a.get_rank() < b.get_rank()
+			else:
+				out = a_marble.get_checkpoint_count() > b_marble.get_checkpoint_count()
 	return out